gcc13 : fix argument format issues
gcc13 based build failed with below error.
  error: static assertion failed: Cannot format an argument.
  To make type T formattable provide a formatter<T> specialization:
  https://fmt.dev/latest/api.html#udt

Problem looks similar to https://github.com/fmtlib/fmt/issues/391

Fix is to convert the "enum" type to "int" in the fmt::format function.

+++ extensions/phal/phal_error.cpp
@@ -417,9 +417,9 @@ void processClockInfoErrorHelper(FFDC* ffdc, const std::string& ffdc_prefix)
 
 void processIplErrorCallback(const ipl_error_info& errInfo)
 {
-    log<level::INFO>(
-        fmt::format("processIplErrorCallback: Error type({})", errInfo.type)
-            .c_str());
+    log<level::INFO>(fmt::format("processIplErrorCallback: Error type({})",
+                                 static_cast<int>(errInfo.type))
+                         .c_str());
 
     switch (errInfo.type)
     {
